unified bright and dark watchfaces
