+++ wear/src/main/AndroidManifest.xml
@@ -28,34 +28,6 @@
         </service>
 
         <service
-            android:name=".HomeDark"
-            android:allowEmbedded="true"
-            android:label="NightWatch(Dark)"
-            android:permission="android.permission.BIND_WALLPAPER">
-            <meta-data android:name="android.service.wallpaper"
-                       android:resource="@xml/watch_face"/>
-            <meta-data android:name="com.google.android.wearable.watchface.preview" android:resource="@drawable/ic_launcher" />
-            <intent-filter>
-                <action android:name="android.service.wallpaper.WallpaperService" />
-                <category android:name="com.google.android.wearable.watchface.category.WATCH_FACE" />
-            </intent-filter>
-        </service>
-
-        <service
-            android:name=".LargeHomeDark"
-            android:allowEmbedded="true"
-            android:label="NightWatch(Large+Dark)"
-            android:permission="android.permission.BIND_WALLPAPER">
-            <meta-data android:name="android.service.wallpaper"
-                       android:resource="@xml/watch_face"/>
-            <meta-data android:name="com.google.android.wearable.watchface.preview" android:resource="@drawable/ic_launcher" />
-            <intent-filter>
-                <action android:name="android.service.wallpaper.WallpaperService" />
-                <category android:name="com.google.android.wearable.watchface.category.WATCH_FACE" />
-            </intent-filter>
-        </service>
-
-        <service
             android:name=".LargeHome"
             android:allowEmbedded="true"
             android:label="NightWatch(Large)"
@@ -70,9 +42,9 @@
         </service>
 
         <service
-            android:name=".ModernWatchfaceDark"
+            android:name=".ModernWatchface"
             android:allowEmbedded="true"
-            android:label="NightWatch(ModernDark)"
+            android:label="NightWatch(Circle)"
             android:permission="android.permission.BIND_WALLPAPER">
             <meta-data android:name="android.service.wallpaper"
                 android:resource="@xml/watch_face"/>
