+++ wear/src/main/java/com/dexdrip/stephenblack/nightwatch/BaseWatchFace.java
@@ -146,8 +146,11 @@ public  abstract class BaseWatchFace extends WatchFace implements SharedPreferen
 
     @Override
     public void onDestroy() {
-        if(localBroadcastManager != null && messageReceiver != null)
-            localBroadcastManager.unregisterReceiver(messageReceiver);
+        if(localBroadcastManager != null && messageReceiver != null){
+            localBroadcastManager.unregisterReceiver(messageReceiver);}
+        if (sharedPrefs != null){
+            sharedPrefs.unregisterOnSharedPreferenceChangeListener(this);
+        }
         super.onDestroy();
     }
 
@@ -230,13 +233,24 @@ public  abstract class BaseWatchFace extends WatchFace implements SharedPreferen
         }
     }
 
-    public void setColor() { Log.e("ERROR: ", "MUST OVERRIDE IN CLASS"); }
+    public void setColor() {
+        if (sharedPrefs.getBoolean("dark", false)) {
+            setColorDark();
+        } else {
+            setColorBright();
+        }
+
+    }
+
+    
 
     @Override
     public void onSharedPreferenceChanged(SharedPreferences sharedPreferences, String key){
         setColor();
         invalidate();
     }
+protected abstract void setColorDark();
+    protected abstract void setColorBright();
 
 
     public void missedReadingAlert() {
