--- wear/src/main/java/com/dexdrip/stephenblack/nightwatch/HomeDark.java
@@ -1,45 +1,0 @@
-package com.dexdrip.stephenblack.nightwatch;
-
-import android.content.Context;
-import android.graphics.Color;
-import android.view.LayoutInflater;
-
-public class HomeDark extends BaseWatchFace {
-    @Override
-    public void onCreate() {
-        super.onCreate();
-        LayoutInflater inflater = (LayoutInflater) getSystemService(Context.LAYOUT_INFLATER_SERVICE);
-        layoutView = inflater.inflate(R.layout.activity_home, null);
-        performViewSetup();
-    }
-
-    @Override
-    public void setColor() {
-        mRelativeLayout.setBackgroundColor(Color.BLACK);
-        mLinearLayout.setBackgroundColor(Color.WHITE);
-        if (sgvLevel == 1) {
-            mSgv.setTextColor(Color.YELLOW);
-            mDirection.setTextColor(Color.YELLOW);
-            mDelta.setTextColor(Color.YELLOW);
-        } else if (sgvLevel == 0) {
-            mSgv.setTextColor(Color.WHITE);
-            mDirection.setTextColor(Color.WHITE);
-            mDelta.setTextColor(Color.WHITE);
-        } else if (sgvLevel == -1) {
-            mSgv.setTextColor(Color.RED);
-            mDirection.setTextColor(Color.RED);
-            mDelta.setTextColor(Color.RED);
-        }
-        if (ageLevel == 1) {
-            mTimestamp.setTextColor(Color.BLACK);
-        } else {
-            mTimestamp.setTextColor(Color.RED);
-        }
-
-        if (batteryLevel == 1) {
-            mUploaderBattery.setTextColor(Color.BLACK);
-        } else {
-            mUploaderBattery.setTextColor(Color.RED);
-        }
-    }
-}
