+++ wear/src/main/java/com/dexdrip/stephenblack/nightwatch/LargeHome.java
@@ -16,11 +16,41 @@ public class LargeHome extends BaseWatchFace {
         performViewSetup();
     }
 
+
     @Override
-    public void setColor() {
+    protected void setColorDark(){
+        mRelativeLayout.setBackgroundColor(Color.BLACK);
+        mLinearLayout.setBackgroundColor(Color.WHITE);
+        if (sgvLevel == 1) {
+            mSgv.setTextColor(Color.YELLOW);
+            mDirection.setTextColor(Color.YELLOW);
+            mDelta.setTextColor(Color.YELLOW);
+        } else if (sgvLevel == 0) {
+            mSgv.setTextColor(Color.WHITE);
+            mDirection.setTextColor(Color.WHITE);
+            mDelta.setTextColor(Color.WHITE);
+        } else if (sgvLevel == -1) {
+            mSgv.setTextColor(Color.RED);
+            mDirection.setTextColor(Color.RED);
+            mDelta.setTextColor(Color.RED);
+        }
+        if (ageLevel == 1) {
+            mTimestamp.setTextColor(Color.BLACK);
+        } else {
+            mTimestamp.setTextColor(Color.RED);
+        }
+
+        if (batteryLevel == 1) {
+            mUploaderBattery.setTextColor(Color.BLACK);
+        } else {
+            mUploaderBattery.setTextColor(Color.RED);
+        }
+    }
 
 
 
+    @Override
+    protected void setColorBright() {
         if (getCurrentWatchMode() == WatchMode.INTERACTIVE) {
             mRelativeLayout.setBackgroundColor(Color.WHITE);
             mLinearLayout.setBackgroundColor(Color.BLACK);
